sorts start (wtf am i doing)

+++ third-sem/LW2/types.h
@@ -10,7 +10,7 @@
 
 struct needle_def {
     int needle; // duh
-    int correct_index; // index where this needle was located
+    long correct_index; // index where this needle was located
     bool should_restore_needle; // false if the needle should not be put into the array after making only one of them
     std::string name;
 };
